backend/requirements.txt
Flask==2.3.3
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.1
google-generativeai==0.4.1
